+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Cli : ajout d'une commande `resolve` permettant de résoudre et tester les chaînes de configuration.
 * Cli : ajout d'une action `--relative-entities` permettant de lister les entités liées à l'entité indiquée.
+* Cli : ajout d'une action `--abort` permettant de d'annuler une exécution de traitement.
 
 ### [Changed]